+++ okta/resources/resources.go
@@ -119,4 +119,6 @@ const (
 	OktaIDaaSUserSchemaProperty            = "okta_user_schema_property"
 	OktaIDaaSUserSecurityQuestions         = "okta_user_security_questions"
 	OktaIDaaSUserType                      = "okta_user_type"
+	OktaGovernanceCampaign                 = "okta_campaign"
+	OktaGovernanceEntitlement              = "okta_entitlement"
 )